backend/requirements.txt
pynest-api==0.4.0
fastapi>=0.115.4
uvicorn[standard]>=0.32.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic>=2.9.2
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
alembic==1.13.0
python-dotenv==1.0.0
qrcode==8.2
pillow==12.0.0

